EventSelectorPopup chooses correct starting index based on given current event name WIP Move CommandSelector to its own PopupWindowContext
