+++ Assets/Fungus/Scripts/Editor/CommandSelectorPopupWindowContent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0e2650550a8549143b9eaab2c2dad511
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
